+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support `timer` and `queue_main` with simple blocks as `Glimmer::LibUI.timer` and `Glimmer::LibUI.queue_main`
 - Support examples/timer.rb
 - Support degrees for arc arguments (instead of radians)
+- Support `circle` shape and use in examples/area_gallery.rb (all versions)
 
 ## 0.1.11